refactor(dispatcher): rename lib/ to src/ for monorepo consistency
