+++ packages/tapeworm_dispatcher_mdb_rmq/index.ts
@@ -1,13 +1,13 @@
-export { Dispatcher } from "./lib/dispatcher";
-export { CommitWatcher } from "./lib/watcher";
-export { CommitPublisher } from "./lib/publisher";
-export { MongoResumeTokenStore } from "./lib/resume/mongodb-store";
+export { Dispatcher } from "./src/dispatcher";
+export { CommitWatcher } from "./src/watcher";
+export { CommitPublisher } from "./src/publisher";
+export { MongoResumeTokenStore } from "./src/resume/mongodb-store";
 
-export type { IResumeTokenStore } from "./lib/resume/types";
+export type { IResumeTokenStore } from "./src/resume/types";
 export type {
   DispatcherConfig,
   DispatcherEvents,
   MongoConfig,
   RabbitConfig,
   ResumeState,
-} from "./lib/types";
+} from "./src/types";
